合并本地和github代码

+++ database/migrations/2020_09_05_134323_create_scores_table.php
@@ -15,7 +15,7 @@ class CreateScoresTable extends Migration
     {
         Schema::create('scores', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->tinyInteger('score')->comment('成绩');
+            $table->unsignedInteger('score')->comment('成绩');
             $table->unsignedBigInteger('student_id')->comment('所属学生');
             $table->unsignedBigInteger('exam_id')->comment('所属考试');
             $table->unsignedBigInteger('course_id')->comment('所属课程');
